jobs/fetch_gsc/requirements.txt
google-api-python-client==2.186.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.38.0
